SLCORE-73 Refactor
