+++ core/src/main/java/org/sonarsource/sonarlint/core/container/connected/update/check/DefaultGlobalStorageUpdateCheckResult.java
@@ -1,0 +1,45 @@
+/*
+ * SonarLint Core - Implementation
+ * Copyright (C) 2009-2016 SonarSource SA
+ * mailto:contact AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonarsource.sonarlint.core.container.connected.update.check;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import org.sonarsource.sonarlint.core.client.api.connected.GlobalStorageUpdateCheckResult;
+
+public class DefaultGlobalStorageUpdateCheckResult implements GlobalStorageUpdateCheckResult {
+
+  private List<String> changelog = new ArrayList<>();
+
+  @Override
+  public boolean needUpdate() {
+    return !changelog.isEmpty();
+  }
+
+  @Override
+  public List<String> changelog() {
+    return Collections.unmodifiableList(changelog);
+  }
+
+  void appendToChangelog(String line) {
+    changelog.add(line);
+  }
+
+}
